requirements.txt
absl-py==2.3.1
certifi==2025.8.3
charset-normalizer==3.4.2
einops==0.8.1
grpcio==1.74.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
lazy_loader==0.4
Markdown==3.8.2
MarkupSafe==3.0.2
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tifffile==2024.8.30
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
